Track view all children orgs with a list
We used to track the state within each org object, which worked when
there's 2 levels. Now that there's 3 levels, updating the org tree means
having to deep copy to the grandchildren level, which the spread
operator cannot do. It feels simpler to track this information using an
external list rather than having it take up space in each org object.